web/I-Sequel/app/requirements.txt
psycopg2-binary==2.9.1
fastapi[all]==0.65.2